+++ libraries/Prism/CHANGELOG.md
@@ -4,7 +4,7 @@ Prism Library Changelog
 ###v1.12
 * Added new third-party libraries - React\Promise, GuzzleHttp\Stream.
 * Upgraded some third-party libraries - Coinbase and GuzzleHttp.
-* Improved helper classes.
+* Improved integration, helper and database classes.
 
 ###v1.11
 * Added command buses that can be used for helper and database collections.